Twitter Sentiment Analysis/Twitter get data/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.2
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.12.1
urllib3==1.26.14
